split terminal.h into window header
